+++ server/controllers/student.controller.js
@@ -1,5 +1,6 @@
 const studentService = require('../services/student.services');
-const studentHelper = require('../helpers/student.helper');
+const commonValidator = require('../models/validators/common.validators');
+const { createStudentFromRequest } = require('../helpers/req2model.helper');
 const { handleApiError } = require('../services/errors/ApiError');
 
 exports.fetchAll = async (req, res) => {
@@ -14,7 +15,14 @@ exports.fetchAll = async (req, res) => {
 
 exports.fetchById = async (req, res) => {
     try{
-        const student = await studentService.fetchById(req.params.studentId);
+        const {studentId} = req.params;
+
+        if(!commonValidator.isValidObjectId(studentId)) {
+            let err = new ApiError(400, 'Hibás azonosító!');
+            return handleApiError(err, res);
+        }
+
+        const student = await studentService.fetchById(studentId);
 
         return res.json(student);
     } catch(error) {
@@ -24,7 +32,7 @@ exports.fetchById = async (req, res) => {
 
 exports.create = async (req, res) => {
     try {
-        const student = studentHelper.createStudentFromRequest(req);
+        const student = createStudentFromRequest(req);
         const newStudent = await studentService.create(student);
 
         return res.json(newStudent);
@@ -37,12 +45,18 @@ exports.update = async (req, res) => {
     try {
         const { studentId } = req.params;
 
+        if(!commonValidator.isValidObjectId(studentId)) {
+            let err = new ApiError(400, 'Hibás azonosító!');
+            return handleApiError(err, res);
+        }
+
         const student = req.body;
 
         const updatedStudent = await studentService.update(studentId, student);
 
         return res.json(updatedStudent);
     } catch (error) {
+        console.log(error);
         return handleApiError(error, res);
     }
 }
@@ -50,6 +64,12 @@ exports.update = async (req, res) => {
 exports.delete = async (req, res) => {
     try {
         const { studentId } = req.params;
+
+        if(!commonValidator.isValidObjectId(studentId)) {
+            let err = new ApiError(400, 'Hibás azonosító!');
+            return handleApiError(err, res);
+        }
+
         const deletedStudent = await studentService.delete(studentId);
 
         return res.json(deletedStudent);
